refactor(oxfmt): Split oxfmt-plugin files (#19031)
Extract main `textToDoc()` implementation.
