+++ apps/oxfmt/src-js/libs/prettier-plugin-oxfmt/text-to-doc.ts
@@ -1,0 +1,35 @@
+import { doc } from "prettier";
+import { format } from "../../index";
+import type { Parser, Doc } from "prettier";
+
+const { hardline, join } = doc.builders;
+const LINE_BREAK_RE = /\r?\n/;
+
+export const textToDoc: Parser<Doc>["parse"] = async (embeddedSourceText, textToDocOptions) => {
+  // NOTE: For (j|t)s-in-xxx, default `parser` is either `babel` or `typescript`
+  const parser = textToDocOptions.parser as string;
+  // In case of ts-in-md, `filepath` is overridden to distinguish TSX or TS
+  const filepath = textToDocOptions.filepath as string;
+  // We need to infer `SourceType` for `oxc_formatter`
+  const filename =
+    parser === "typescript"
+      ? filepath.endsWith(".tsx")
+        ? "dummy.tsx" // tsx-in-md
+        : "dummy.ts" // ts-in-md / ts-in-xxx
+      : "dummy.jsx"; // Otherwise, always enable JSX for js-in-xxx, it's safe
+
+  // NOTE: Ultimately, this should be `textToDoc()` like Prettier originally does
+  const { code, errors } = await format(
+    filename,
+    embeddedSourceText,
+    // SAFETY: This is generated by Rust side and only available if plugin is used
+    JSON.parse(textToDocOptions._oxfmtPluginOptionsJson as string),
+  );
+
+  if (0 < errors.length) throw new Error(errors[0].message);
+
+  // NOTE: This is required for the parent ((j|t)s-in-xxx) printer
+  // to handle line breaks correctly,
+  // not only for `options.vueIndentScriptAndStyle` but also for basic printing.
+  return join(hardline, code.split(LINE_BREAK_RE));
+};
